+++ extension/src/utils/constants.ts
@@ -1,0 +1,2 @@
+export const REFRESH_PATH_POWERSHELL = '$env:Path = [System.Environment]::GetEnvironmentVariable("Path","Machine") + ";" + [System.Environment]::GetEnvironmentVariable("Path","User");'
+export const FILE_PATH_EMPTY = ''
