+++ extension/src/utils/utils.ts
@@ -1,9 +1,5 @@
-import { window } from 'vscode'
-import { execDefault, execPowerShell } from './exec-shell'
-
-export const FILE_PATH_EMPTY = ''
-
-const REFRESH_PATH_POWERSHELL = '$env:Path = [System.Environment]::GetEnvironmentVariable("Path","Machine") + ";" + [System.Environment]::GetEnvironmentVariable("Path","User");'
+import { execDefault, execPowerShell } from './exec-system'
+import { REFRESH_PATH_POWERSHELL } from './constants'
 
 // Check if a promise was resolved or rejected
 export async function didResolve (promise: Promise<any>): Promise<boolean> {
@@ -13,43 +9,6 @@ export async function didResolve (promise: Promise<any>): Promise<boolean> {
   } catch (e) {
     return false
   }
-}
-
-// Execute a command in vscode terminal
-export async function execVscodeTerminal (name: string, command: string, shellPath?: string): Promise<void> {
-  const OS_TYPE = process.platform
-  if (shellPath == null) { shellPath = OS_TYPE ? 'powershell.exe' : '/bin/bash' }
-
-  const term = window.createTerminal({
-    name,
-    hideFromUser: false,
-    shellPath
-  })
-
-  // Must refresh path in windows
-  if (OS_TYPE === 'win32') { command = REFRESH_PATH_POWERSHELL + command }
-
-  // Add exit to command to close terminal
-  command = OS_TYPE === 'win32' ? command + '; exit $LASTEXITCODE' : command + '; exit $?'
-
-  term.sendText(command)
-  term.show()
-
-  // Wait for installation to complete
-  await new Promise<void>((resolve, reject) => {
-    const disposeToken = window.onDidCloseTerminal(
-      async (closedTerminal) => {
-        if (closedTerminal === term) {
-          disposeToken.dispose()
-          if (term.exitStatus?.code === 0) {
-            resolve()
-          } else {
-            reject('Terminal execution failed')
-          }
-        }
-      }
-    )
-  })
 }
 
 export async function delay (seconds: number): Promise<void> {
